pyproject.toml
[tool.poetry]
name = "starry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
pytz = "^2020.1"
async-timeout = "^3.0.1"
flask = "^1.1.2"
lavalink = "^3.1.2"
ffmpeg = "^1.4"
python-ffmpeg = "^1.0.11"
youtube_dl = "^2020.7.28"
uwuify = "^1.0.1"
"nekos.py" = "^1.0.3"
"discord.py" = "^1.4.1"
pymongo = "^3.11.0"
dnspython = "^2.0.0"
discord-webhook = "^0.8.0"
tweepy = "^3.9.0"
db = "^0.1.1"
utils = "^1.0.1"
monotonic = "^1.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
